microbit: Update of maximum BLE bond count from 2 to 4.
BLE bond information is stored in a table in flash memory during the pairing process with another BLE device.
This information is required for a secure BLE connection.

There is a limit to the number of entries in this table. It used to be 4 but was recently changed to 2
during a code merge with mbed updates. This patch returns to support four bonds.